Upload Visualizations
Uploaded you will find two different ways of visualizing the data from RetroSheets.org and the Baseball database Josh Luedeman and I have created.

Each will show you the best day of the week to see your favorite baseball team play based on historical data.  This is not a definite prediction of future success and should be used for fun or learning.